feito incremento final

+++ app/Services/UserService.php
@@ -4,13 +4,19 @@ namespace App\Services;
 
 use App\Models\Movie;
 use App\Models\User;
+use App\Models\Vote;
 use App\Repository\MovieRepo;
+use App\Repository\UserRepo;
 use Exception;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
+use PHPOpenSourceSaver\JWTAuth\Facades\JWTAuth;
 
 class UserService
 {
 
-    public function create(array $request): array
+    public function register(array $request): array
     {
         try {
             $returnValidate = ValidateService::paramsCreateUser($request);
@@ -18,13 +24,53 @@ class UserService
                 throw new Exception($returnValidate['message']);
             }
 
-            if (User::create($returnValidate['params']) === false) {
-                throw new Exception("Erro ao salvar Usuário. Por favor refaça o procedimento");
-            }
+            $user = UserRepo::create($returnValidate['params']);
 
+            $token = JWTAuth::fromUser($user);
             return [
                 'error' => false,
-                'message' => 'Usuário cadastrado com sucesso'
+                'message' => 'Usuário cadastrado com sucesso',
+                'user' => [
+                    'id' => $user->id,
+                    'name' => $user->name,
+                    'group' => $user->group_user_id,
+                ],
+                'authorization' => [
+                    'token' => $token,
+                    'type' => 'bearer',
+                ]
+            ];
+        } catch (\Throwable $th) {
+            return [
+                'error' => true,
+                'message' => $th->getMessage()
+            ];
+        }
+    }
+
+    public function login(array $params): array
+    {
+        try {
+            // $credentials = $request->only('email', 'password');
+            $token = JWTAuth::attempt($params);
+
+            if (!$token) {
+                throw new Exception("Não autorizado");
+            }
+
+            $user = Auth::user();
+            return [
+                'error' => false,
+                'message' => 'Usuário autenticado com sucesso',
+                'user' => [
+                    'id' => $user->id,
+                    'name' => $user->name,
+                    'group' => $user->group_user_id
+                ],
+                'authorization' => [
+                    'token' => $token,
+                    'type' => 'bearer',
+                ]
             ];
         } catch (\Throwable $th) {
             return [
@@ -78,13 +124,40 @@ class UserService
                 throw new Exception("Usuário inexistente");
             }
 
+            DB::beginTransaction();
+            if (Vote::where('users_id', $id)->delete() === false) {
+                throw new Exception("Erro ao deletar os votos do usuários");
+            }
+
             if (!User::destroy($id)) {
                 throw new Exception("Erro ao deletar Usuário");
             }
+            DB::commit();
 
             return [
                 'error' => false,
                 'message' => 'Usuário deletado com sucesso'
+            ];
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            return [
+                'error' => true,
+                'message' => $th->getMessage()
+            ];
+        }
+    }
+
+    public function show()
+    {
+        try {
+            $users = UserRepo::show();
+            if ($users === false) {
+                throw new Exception("Erro ao buscar os Usuários. Por favor tente mais tarde");
+            }
+
+            return [
+                'error' => false,
+                'message' => count($users) ? $users : 'Nenhum Usuário cadastrado'
             ];
         } catch (\Throwable $th) {
             return [
